Exported function for creating metric logic docs. Fixed unit tests.
