+++ src/index.ts
@@ -18,6 +18,7 @@ import {
   cleanMetricComputations,
   cleanMetricExecutions,
   createMetricComputation,
+  createMetricLogicDoc,
   delayFormSubmissionAndCheckIfCancelled,
   distribute,
   distributeLater,
@@ -34,7 +35,7 @@ import {createViewLogicFn, onMessageViewLogicsQueue} from "./logics/view-logics"
 import {resetUsageStats, stopBillingIfBudgetExceeded, useBillProtect} from "./utils/bill-protect";
 import {Firestore} from "firebase-admin/firestore";
 import {DatabaseEvent, DataSnapshot, onValueCreated} from "firebase-functions/v2/database";
-import {_mockable as _pathMockable, parseEntity} from "./utils/paths";
+import {parseEntity} from "./utils/paths";
 import {database} from "firebase-admin";
 import {initClient} from "emberflow-admin-client/lib";
 import {internalDbStructure, InternalEntity} from "./db-structure";
@@ -152,15 +153,7 @@ export function initializeEmberFlow(
   const logicNames = logicConfigs.map((config) => config.name);
   const viewLogicNames = viewLogicConfigs.map((config) => config.name);
   const allLogicNames = [...logicNames, ...viewLogicNames];
-  allLogicNames.forEach(async (logicName) => {
-    const metricsRef = db.doc(`@metrics/${logicName}`);
-    if (!await _pathMockable.doesPathExists(metricsRef.path)) {
-      await metricsRef.set({
-        totalExecTime: 0,
-        totalExecCount: 0,
-      });
-    }
-  });
+  allLogicNames.forEach(createMetricLogicDoc);
 
   functionsConfig["onFormSubmit"] = onValueCreated(
     {
