+++ src/tests/index-utils.test.ts
@@ -34,6 +34,9 @@ import {
 } from "../index-utils";
 import * as viewLogics from "../logics/view-logics";
 
+// should mock when using initializeEmberFlow and testing db.doc() calls count
+jest.spyOn(indexUtils, "createMetricLogicDoc").mockResolvedValue();
+
 jest.spyOn(console, "log").mockImplementation();
 jest.spyOn(console, "info").mockImplementation();
